requirements.txt
langchain
openai
python_dotenv
requests
newspaper3k